split jglue.py to task-specific file
